requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.17
pydantic>=2.6.0
pydantic-settings>=2.1.0
alembic>=1.13.1
openpyxl>=3.1.2
fpdf2>=2.8.0
lxml>=5.1.0
cryptography>=42.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
celery>=5.3.0
redis>=5.0.0
